+++ pyvista/core/utilities/geometric_sources.py
@@ -14,7 +14,6 @@ from typing import List
 from typing import Sequence
 from typing import Tuple
 from typing import Union
-import weakref
 
 import numpy as np
 from vtkmodules.vtkRenderingFreeType import vtkVectorText
@@ -908,17 +907,6 @@ class Text3DSource(vtkVectorText):
         """Initialize source."""
         super().__init__()
 
-        # Create output filters to make text 3D
-        extrude = _vtk.vtkLinearExtrusionFilter()
-        extrude.SetInputConnection(self.GetOutputPort())
-        extrude.SetExtrusionTypeToNormalExtrusion()
-        extrude.SetVector(0, 0, 1)
-        self._extrude_filter = weakref.ref(extrude)
-
-        tri_filter = _vtk.vtkTriangleFilter()
-        tri_filter.SetInputConnection(extrude.GetOutputPort())
-        self._tri_filter = weakref.ref(tri_filter)
-
         self._output = pyvista.PolyData()
 
         # Set params
@@ -1049,8 +1037,16 @@ class Text3DSource(vtkVectorText):
                 out = self.GetOutput()
             else:
                 # 3D case, apply filters
-                self._tri_filter.Update()
-                out = self._tri_filter.GetOutput()
+                # Create output filters to make text 3D
+                extrude = _vtk.vtkLinearExtrusionFilter()
+                extrude.SetInputConnection(self.GetOutputPort())
+                extrude.SetExtrusionTypeToNormalExtrusion()
+                extrude.SetVector(0, 0, 1)
+
+                tri_filter = _vtk.vtkTriangleFilter()
+                tri_filter.SetInputConnection(extrude.GetOutputPort())
+                tri_filter.Update()
+                out = tri_filter.GetOutput()
 
             # Modify output object
             self._output.copy_from(out)
